function/requirements.txt
google-cloud-bigquery==3.2.0
pyarrow==6.0.1
numpy==1.21.4

